[nrf51822][gcc_arm]: fix makefile templates and nrf_svc.h's 'impossible constraint in asm' error

+++ libraries/mbed/targets/hal/TARGET_NORDIC/TARGET_MCU_NRF51822/Lib/s110_nrf51822_7_0_0/s110_nrf51822_7.0.0_API/include/nrf_svc.h
@@ -16,7 +16,7 @@
   { \
     __asm( \
         "svc %0\n" \
-        "bx r14" : : "I" (number) : "r0" \
+        "bx r14" : : "I" ((uint32_t)number) : "r0" \
     ); \
   }
 #elif defined (__ICCARM__)
